Move pkg/quickstart to pkg/gitops/profile
