+++ pkg/gitops/profile/url_test.go
@@ -1,11 +1,11 @@
-package quickstart_test
+package profile_test
 
 import (
 	"testing"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/weaveworks/eksctl/pkg/quickstart"
+	"github.com/weaveworks/eksctl/pkg/gitops/profile"
 	"github.com/weaveworks/eksctl/pkg/testutils"
 )
 
@@ -13,10 +13,10 @@ func TestSuite(t *testing.T) {
 	testutils.RegisterAndRun(t)
 }
 
-var _ = Describe("quickstart", func() {
+var _ = Describe("profile", func() {
 	Describe("RepositoryURL", func() {
 		It("returns Git URLs as-is", func() {
-			url, err := quickstart.RepositoryURL("https://github.com/eksctl-bot/my-gitops-repo")
+			url, err := profile.RepositoryURL("https://github.com/eksctl-bot/my-gitops-repo")
 			Expect(err).To(Not(HaveOccurred()))
 			Expect(url).To(Equal("https://github.com/eksctl-bot/my-gitops-repo"))
 		})
@@ -30,14 +30,14 @@ var _ = Describe("quickstart", func() {
 				{mnemonic: "appmesh", url: "https://github.com/weaveworks/eks-appmesh-profile"},
 			}
 			for _, mnemonicToURL := range mnemonicToURLs {
-				url, err := quickstart.RepositoryURL(mnemonicToURL.mnemonic)
+				url, err := profile.RepositoryURL(mnemonicToURL.mnemonic)
 				Expect(err).To(Not(HaveOccurred()))
 				Expect(url).To(Equal(mnemonicToURL.url))
 			}
 		})
 
 		It("returns an error otherwise", func() {
-			url, err := quickstart.RepositoryURL("foo")
+			url, err := profile.RepositoryURL("foo")
 			Expect(err).To(HaveOccurred())
 			Expect(err.Error()).To(Equal("invalid URL or unknown Quick Start profile: foo"))
 			Expect(url).To(Equal(""))
